Reset checker state (movement options) whenever a checker is moved.
This code update is mainly comments. There aren't very many new lines of code actually added. This will help eliminate some of the out of bounds errors that you were getting. I did not indicate that this was something you would need to take care of in the suggested order that I gave to you for coding the rest of the program, and so I felt it best to give this to you.